pwa now works, and omg why is the author an idiot
Ok i need to vent. Why on earth does a plugin designed for vite and easily building PWAs not auto register the sw, nor add manifest link! WTF!! In the example it adds it through a script manually editing the html, which is never mentioned anywhere, and you have to hunt for this based on an offhand mention!!